+++ src/components/Main/Main.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 import { Route, Routes } from 'react-router-dom'
+import axios from 'axios'
 import Form from '../Form'
 import Home from '../Home'
 import NewsList from '../NewsList'
@@ -17,27 +18,30 @@ export default class Main extends Component {
   createNew = (article) => {
     this.setState({NewsList: [...this.state.NewsList, ...article]})
   }
-
-  handleSubmit = (event) => {
-    event.preventDefault()
-
-    const article = {
-      title: event.target.title.value,
-      text: event.target.text.value,
-      date: event.target.date.value,
-      category: event.target.category.value,
-      image: event.target.image.value,
-      url: event.target.url.value
-    }
-
-    this.createNew([article])
-  }
-
-  deleteAllINews = () => this.setState({ NewsList: [] })
     
   deleteNew = i => {
       const NewsList = this.state.NewsList.filter((article, j) => j !== i)
       this.setState({ NewsList })
+  }
+
+  async componentDidMount() {
+    const resp = await axios.get(`https://api.nytimes.com/svc/topstories/v2/technology.json?api-key=${process.env.REACT_APP_API_KEY}`)
+    const data = resp.data
+    
+    const newsArray = data.results.map( element => {
+                        return {
+                          title: element.title,
+                          text: element.abstract,
+                          date: element.published_date,
+                          category: element.section,
+                          image: element.multimedia[0].url,
+                          url: element.short_url
+                        }
+    })
+
+    const newsList = newsArray.slice(0, 5)
+  
+    this.createNew(newsList);
   }
 
   render() {
@@ -45,8 +49,8 @@ export default class Main extends Component {
       <main>
         <Routes>
           <Route path="/" element={<Home/>} exact />  
-          <Route path="/form" element={<Form submit={this.handleSubmit}/>} /> 
-          <Route path="/list" element={<NewsList newsList={this.state.NewsList} createNew={this.createNew} deleteNew={this.deleteNew}/>} /> 
+          <Route path="/form" element={<Form createNew={this.createNew}/>} /> 
+          <Route path="/list" element={<NewsList newsList={this.state.NewsList} deleteNew={this.deleteNew}/>} /> 
         </Routes>
       </main>
     )
